agents/specialized/travel_agent/requirements.txt
# Core dependencies
pydantic>=2.10.4,<3.0.0
browser-use==0.1.37
playwright>=1.50.0
aiohttp>=3.11.12
beautifulsoup4>=4.13.3
python-dotenv>=1.0.1

# Database and storage
supabase>=2.3.0
pandas>=2.0.0
numpy>=1.26.4
pgvector>=0.2.5

# GPT and LangChain integration
openai>=1.12.0
langchain>=0.3.14
langchain-openai>=0.3.1
langsmith>=0.1.17,<0.3

# gRPC dependencies
grpcio<1.68.0
grpcio-status<1.68.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
pytest-mock>=3.12.0
